refactor(map): convert view-builder part files to standalone classes (#3588)
Replace the `part`/`part of` coupling of the init_game_map_view_builder
family with standalone, importable classes:

- `InitGameMapViewMarkers` (fleet/capital/port/town/warp/sea-zone markers)
- `InitGameMapViewCells` (cell-grid + unit/civilian marker overlays)

The orchestration entrypoint `buildInitGameMapViewData` stays in
`init_game_map_view_builder.dart` and now imports the two classes instead
of relying on a shared library scope. Behavior is unchanged; the three
`*_part.dart` fragments are deleted.

Refs #3588
